+++ libemulation/Implementation/Generic/ROM.cpp
@@ -44,7 +44,7 @@ bool ROM::init()
 		return false;
 	}
 	
-	int size = getNextPowerOf2(data.size());
+	int size = getNextPowerOf2((int) data.size());
 	data.resize(size);
 	datap = (OEUInt8 *) &data.front();
 	mask = size - 1;
